+++ src/parser/block/extension/figbase.test.ts
@@ -1,0 +1,28 @@
+import { figbase } from './figbase';
+import { some } from '../../../combinator';
+import { inspect } from '../../../debug.test';
+
+describe('Unit: parser/block/extension/figbase', () => {
+  describe('figbase', () => {
+    const parser = (source: string) => some(figbase)(source, {});
+
+    it('invalid', () => {
+      assert.deepStrictEqual(inspect(parser('\n$-0')), undefined);
+      assert.deepStrictEqual(inspect(parser('$-0.')), undefined);
+      assert.deepStrictEqual(inspect(parser('$-0\n 0')), undefined);
+      assert.deepStrictEqual(inspect(parser('$-name')), undefined);
+      assert.deepStrictEqual(inspect(parser('$-name-0')), undefined);
+      assert.deepStrictEqual(inspect(parser('$group-0')), undefined);
+      assert.deepStrictEqual(inspect(parser(' [$-0]')), undefined);
+    });
+
+    it('valid', () => {
+      assert.deepStrictEqual(inspect(parser('$-0')), [['<figure data-label="$-0" data-group="$" style="display: none;"></figure>'], '']);
+      assert.deepStrictEqual(inspect(parser('$-0.0')), [['<figure data-label="$-0.0" data-group="$" style="display: none;"></figure>'], '']);
+      assert.deepStrictEqual(inspect(parser('$-0\n \n')), [['<figure data-label="$-0" data-group="$" style="display: none;"></figure>'], ' \n']);
+      assert.deepStrictEqual(inspect(parser('[$-0]')), [['<figure data-label="$-0" data-group="$" style="display: none;"></figure>'], '']);
+    });
+
+  });
+
+});